requirements.txt
accelerate==1.8.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
airportsdata==20250706
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.46.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
asyncpg==0.30.0
attrs==25.3.0
authlib==1.6.1
av==15.0.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blake3==1.0.5
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.9
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.16
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
cmake==4.0.3
coloredlogs==15.0.1
compressed-tensors==0.10.2
conda-pack==0.8.1
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.5
cupy-cuda12x==13.5.1
cycler==0.12.1
cyclopts==3.22.3
dataclasses-json==0.6.7
datasets==1.18.3
deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0
docutils==0.21.2
docx==0.2.4
duckduckgo-search==8.1.1
durationpy==0.10
dydantic==0.0.8
effdet==0.4.1
einops==0.8.1
email-validator==2.2.0
emoji==2.14.1
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastmcp==2.10.6
fastrlock==0.8.3
filelock==3.18.0
filetype==1.2.0
flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gguf==0.17.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
google-genai==1.26.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hanlp-downloader==0.0.25
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.12
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
interegular==0.3.3
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
jpype1==1.6.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-pydantic==0.6
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.4
kubernetes==33.1.0
langchain==0.3.26
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.72
langchain-experimental==0.3.4
langchain-mcp-tools==0.2.10
langchain-ollama==0.3.4
langchain-openai==0.3.28
langchain-postgres==0.0.15
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langdetect==1.0.9
langgraph==0.6.1
langgraph-checkpoint==2.1.0
langgraph-checkpoint-postgres==2.0.23
langgraph-prebuilt==0.6.1
langgraph-sdk==0.2.0
langmem==0.0.29
langsmith==0.4.4
lark==1.2.2
llama-cpp-python==0.3.14
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==6.0.0
markdown-it-py==3.0.0
markdown2==2.5.3
markdownify==1.1.0
marker-pdf==1.8.2
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mcp==1.12.2
mdurl==0.1.2
mistral-common==1.7.0
mmh3==5.2.0
modelscope==1.28.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
olefile==0.47
ollama==0.5.1
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.19.2
openai==1.90.0
openapi-pydantic==0.5.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.18
ormsgpack==1.10.0
oss2==2.19.1
outlines==0.1.11
outlines-core==0.1.26
overrides==7.7.0
packaging==24.2
pandas==2.3.1
partial-json-parser==0.2.1.1.post6
pdf2image==1.17.0
pdfminer-six==20250506
pdfplumber==0.11.7
pdftext==0.6.3
pgvector==0.3.6
pi-heif==1.0.0
pikepdf==9.10.2
pillow==10.4.0
platformdirs==4.3.8
portalocker==3.2.0
posthog==5.4.0
pre-commit==4.2.0
primp==0.15.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.1
pycocotools==2.0.10
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyhanlp==0.1.89
pyparsing==3.2.3
pypdf==5.8.0
pypdf2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
qdrant-client==1.15.0
qwen-vl-utils==0.0.11
rapidfuzz==3.13.0
ray==2.47.1
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.18.0
rich==14.0.0
rich-rst==1.3.1
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.1
surya-ocr==0.14.6
sympy==1.14.0
tavily-python==0.7.10
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.17
tokenizers==0.21.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.53.1
triton==3.3.0
trustcall==0.0.39
typer==0.16.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unstructured==0.18.9
unstructured-client==0.39.1
unstructured-inference==1.0.5
unstructured-pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.32.0
vllm==0.9.2
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.30
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
prefix: /opt/anaconda3/envs/langchain
